+++ axelor-account/src/main/java/com/axelor/apps/account/report/ITranslation.java
@@ -433,4 +433,28 @@ public interface ITranslation {
       /*$$(*/ "AccountingReportType17.year_start_value"; /*)*/
   public static final String ACCOUNTING_REPORT_17_DEPRECIATION_CHARGE =
       /*$$(*/ "AccountingReportType17.depreciation_charge"; /*)*/
+
+  public static final String PAYMENT_REMINDER_PAYMENT_REMINDER =
+      /*$$(*/ "PaymentReminder.payment_reminder"; /*)*/
+  public static final String PAYMENT_REMINDER_REFERENCE_DATE =
+      /*$$(*/ "PaymentReminder.reference_date"; /*)*/
+  public static final String PAYMENT_REMINDER_LAST_REMINDER_DATE =
+      /*$$(*/ "PaymentReminder.last_reminder_date"; /*)*/
+  public static final String PAYMENT_REMINDER_CURRENCY_CODE =
+      /*$$(*/ "PaymentReminder.currency_code"; /*)*/
+  public static final String PAYMENT_REMINDER_IDENTIFICATION_CODE =
+      /*$$(*/ "PaymentReminder.identification_code"; /*)*/
+  public static final String PAYMENT_REMINDER_DATE = /*$$(*/ "PaymentReminder.date"; /*)*/
+  public static final String PAYMENT_REMINDER_LATE_PAYMMENT_DAYS =
+      /*$$(*/ "PaymentReminder.late_payment_days"; /*)*/
+  public static final String PAYMENT_REMINDER_AMOUNT = /*$$(*/ "PaymentReminder.amount"; /*)*/
+  public static final String PAYMENT_REMINDER_AMOUNT_PAID =
+      /*$$(*/ "PaymentReminder.amount_paid"; /*)*/
+  public static final String PAYMENT_REMINDER_TO_PAY = /*$$(*/ "PaymentReminder.to_pay"; /*)*/
+  public static final String PAYMENT_REMINDER_TOTAL_AMOUNT =
+      /*$$(*/ "PaymentReminder.total_amount"; /*)*/
+  public static final String PAYMENT_REMINDER_TOTAL_AMOUNT_PAID =
+      /*$$(*/ "PaymentReminder.total_amount_paid"; /*)*/
+  public static final String PAYMENT_REMINDER_TOTAL_TO_PAY =
+      /*$$(*/ "PaymentReminder.total_to_pay"; /*)*/
 }